refactor: extract StreamingFileImportService partials by concern (#1144)
Main file 3,025 → 583 LOC (-81%). Partials by concern:
- Batch.cs / Streaming.cs / Preview.cs (chunked read/preview pipeline)
- CleanupScratch.cs / PrepareScratch.cs (scratch lifecycle)
- CrsDetection.cs / GeoParquetPrep.cs (format prep)
- Logging.cs (source-gen LoggerMessage)
- Readers.cs / Streams.cs (format-specific I/O)
- Sql.cs (DDL/DML generation)
- Validation.cs (input validation)
