+++ src/Honua.Postgres/Features/Import/StreamingFileImportService.Sql.cs
@@ -1,0 +1,63 @@
+// Copyright (c) Honua. All rights reserved.
+// Licensed under the Elastic License 2.0. See LICENSE in the project root.
+
+using Honua.Postgres.Features.Infrastructure;
+using Npgsql;
+
+namespace Honua.Postgres.Features.Import;
+
+internal sealed partial class StreamingFileImportService
+{
+    private static string QuoteIdentifier(string identifier)
+    {
+        return $"\"{identifier.Replace("\"", "\"\"")}\"";
+    }
+
+    private static string GetAllowedTableName(string tableName)
+    {
+        ValidateTableName(tableName);
+        var sanitized = System.Text.RegularExpressions.Regex.Replace(tableName, @"[^a-zA-Z0-9_]", "_");
+        return "imported_" + sanitized.ToLowerInvariant();
+    }
+
+    private static async Task CreateTableAsync(
+        NpgsqlConnection connection,
+        string schemaName,
+        string tableName,
+        int targetSrid,
+        CancellationToken cancellationToken)
+    {
+        await using var command = new NpgsqlCommand(CreateImportTableSql, connection);
+        command.Parameters.AddWithValue("schema_name", schemaName);
+        command.Parameters.AddWithValue("table_name", tableName);
+        command.Parameters.AddWithValue("target_srid", targetSrid);
+        await command.ExecuteNonQueryAsync(cancellationToken);
+    }
+
+    private static async Task AnalyzeTableAsync(
+        NpgsqlConnection connection,
+        string schemaName,
+        string tableName,
+        CancellationToken cancellationToken)
+    {
+        await using var command = PostgresSqlSafety.CreateAnalyzeCommand(connection, schemaName, tableName);
+        await command.ExecuteNonQueryAsync(cancellationToken);
+    }
+
+    private string ResolveTargetSchema(string? requestedSchema)
+    {
+        var schema = string.IsNullOrWhiteSpace(requestedSchema)
+            ? _schemaConfiguration.DefaultOperationalSchema
+            : requestedSchema.Trim();
+
+        if (!SchemaSearchPath.IsValidIdentifier(schema))
+        {
+            throw new ArgumentException("Target schema contains invalid characters.", nameof(requestedSchema));
+        }
+
+        return schema;
+    }
+
+    private Task<NpgsqlConnectionLease> OpenConnectionAsync(CancellationToken cancellationToken)
+        => _connectionProvider.OpenNpgsqlConnectionAsync(cancellationToken);
+}
